Refactor: AddButton base component
Add ingredient to RecipeBoard
