+++ planner-ui/src/components/AddButton.tsx
@@ -1,0 +1,32 @@
+import { Button } from "@mui/material";
+import { useState, FC, ComponentType } from "react";
+import PostAddIcon from "@mui/icons-material/PostAdd";
+import { AddModalProps } from "../interfaces/AddModalProps";
+
+interface AddButtonProps {
+  label: string;
+  Modal: ComponentType<AddModalProps>;
+}
+
+const AddButton: FC<AddButtonProps> = ({ label, Modal }) => {
+  const [openModal, setOpenModal] = useState(false);
+  const handleOpen = () => setOpenModal(true);
+  const handleClose = () => {
+    setOpenModal(false);
+  };
+
+  return (
+    <div>
+      <Button
+        aria-label={`${label} button`}
+        onClick={handleOpen}
+        endIcon={<PostAddIcon />}
+      >
+        {label}
+      </Button>
+      <Modal open={openModal} handleClose={handleClose} />
+    </div>
+  );
+};
+
+export default AddButton;
