+++ planner-ui/src/interfaces/AddModalProps.ts
@@ -1,0 +1,4 @@
+export interface AddModalProps {
+  open: boolean;
+  handleClose: () => void;
+}
